+++ cli/create-seedcord/src/interview/steps/directory.ts
@@ -4,7 +4,7 @@ import { SeedcordError } from '@seedcord/errors/internal';
 
 import { requireAnswer } from './requireAnswer';
 
-import type { Step } from '../types';
+import type { Step } from '@interview/types';
 
 function parseDirectory(raw: string): string {
     const value = raw.trim();
