+++ cli/create-seedcord/src/interview/steps/index.ts
@@ -6,7 +6,7 @@ import { publicKeyStep } from './publicKey';
 import { tokenStep } from './token';
 import { transportStep } from './transport';
 
-import type { AnyStep } from '../types';
+import type { AnyStep } from '@interview/types';
 
 // the token and public key are both read off the Discord dashboard
 export const STEPS: AnyStep[] = [
