+++ cli/create-seedcord/tests/interview/colorNames.types.test.ts
@@ -1,7 +1,7 @@
 import { describe, expectTypeOf, it } from 'vitest';
 
+import type { ColorChoice } from '@interview/steps/botColor';
 import type { ColorName } from '@seedcord/types';
-import type { ColorChoice } from '@src/interview/steps/botColor';
 
 // the `satisfies ColorName[]` on the list already proves the other direction
 describe('the color names the picker offers', () => {
