+++ cli/create-seedcord/tsconfig.json
@@ -5,7 +5,10 @@
         "rootDir": "./",
         "outDir": "./dist",
         "paths": {
-            "@src/*": ["./src/*"]
+            "@cli/*": ["./src/cli/*"],
+            "@interview/*": ["./src/interview/*"],
+            "@scaffold/*": ["./src/scaffold/*"],
+            "@template/*": ["./src/template/*"]
         }
     },
     "include": ["src/**/*.ts", "tests/**/*.ts", "tsdown.config.ts", "vitest.config.ts"],
